Bugfix
Data size was too small for skill efficiencies and likely overflowing for all but the smallest test cases.

+++ CrafterOfLight/Player.cpp
@@ -182,7 +182,7 @@ void Player::SkillEffect() {
 
 void Player::SynthesisBuffs() {
 	currentSkill.synthesisEfficiency *= progressPerOneEfficiency;
-	const uint8_t efficiency = currentSkill.synthesisEfficiency;
+	const uint16_t efficiency = currentSkill.synthesisEfficiency;
 	if (currentPlayerState.buffs[MUSCLEMEMORY] > 0) {
 		currentPlayerState.buffs[MUSCLEMEMORY] = 0;
 		currentSkill.synthesisEfficiency += efficiency;
@@ -197,7 +197,7 @@ void Player::SynthesisBuffs() {
 void Player::TouchBuffs(uint8_t innerQuietStacks) {
 	ApplyInnerQuiet();
 	currentSkill.touchEfficiency *= qualityPerOneEfficiency;
-	const uint8_t efficiency = currentSkill.touchEfficiency;
+	const uint16_t efficiency = currentSkill.touchEfficiency;
 	if (currentPlayerState.buffs[GREATSTRIDES] > 0) {
 		currentPlayerState.buffs[GREATSTRIDES] = 0;
 		currentSkill.touchEfficiency += efficiency;
